requirements.txt
psycopg2-binary
sqlalchemy
google-generativeai
python-dotenv
PyYAML